Refactor THPUtils_invalidArguments into separate file
